[CORDA-1970]: Refactored service layer to use clojures.
